requirements.txt
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1